+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
 ### Added
 - Computation of the norm
+- Optional handling in or [PR6](https://github.com/Qrlew/qrlew/pull/6)
 
 ## [0.1.2] - 2023-06-01
 ### Added